requirements.txt
pytest==8.3.5
pytest-playwright==0.6.2
pytest-html==4.1.1
python-dotenv==1.1.0
